+++ crates/module-system/module-implementations/sov-evm/tests/integration/main.rs
@@ -3,6 +3,7 @@ mod config;
 mod contracts;
 mod genesis;
 mod helpers;
+mod max_fee;
 mod pruning;
 mod runtime;
 mod state;